SWMM-HEAT Integration in SWMM GUI
SWMM-HEAT integrated for the most part with the original SWMM GUI, now it can assign and edit direct and dry weather inflows, edit the WTEMPERATURE object under Water Quality, run the simulation, read binary output,  generate tables and graphs. Has an option to enable or disable the temperature model in the options.